feat: Enhance game state management and pattern validation
- Introduced game state enumeration for better state handling.
- Added functions to retrieve game state and cooldown/session timings.
- Updated pattern structure to use uint16 for pattern count.
- Improved JSON pattern validation with new constraints on IDs, names, and weights.
- Added new patterns and validation rules for laser pulse duration and behavior.
- Implemented CI workflow for automated testing and validation of patterns.
- Documented new architecture and runtime limits for patterns in v2 documentation.

+++ src/app_util.h
@@ -6,7 +6,7 @@
 #include "main.h"
 #include "freertos/FreeRTOS.h"
 
-#define ARRAY_SIZE(a) ((uint8_t)(sizeof(a) / sizeof((a)[0])))
+#define ARRAY_SIZE(a) (sizeof(a) / sizeof((a)[0]))
 
 static inline portTickType ms_to_ticks_min1(uint32_t ms)
 {
